Finalizando simulador

+++ domain/Intervalo.java
@@ -1,19 +1,19 @@
 package domain;
 
 public class Intervalo {
-    private final int inicio;
-    private final int fim;
+    private final double inicio;
+    private final double fim;
 
-    public Intervalo(int inicio, int fim) {
+    public Intervalo(double inicio, double fim) {
         this.inicio = inicio;
         this.fim = fim;
     }
 
-    public int getInicio() {
+    public double getInicio() {
         return inicio;
     }
 
-    public int getFim() {
+    public double getFim() {
         return fim;
     }
 }
